added helpers.py, system-test.py
