+++ system-test.py
@@ -1,0 +1,12 @@
+import helpers
+
+if __name__ == '__main__':
+
+    rpc_client = helpers.RpcClient('predictor-preprocessor')
+
+    with open('test_assets/example.json', 'r') as f:
+        contents = f.read()
+
+    result = rpc_client.call(contents)
+    
+    print(result)
